insert nav component and insert house icon
